Flatten directory structure
Components, theme, Storybook, and unit test.
