+++ components/SignOutButton.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import { signOut } from "next-auth/react"
-import Button from "../ui/button/Button"
+import Button from "./Button"
 
 export default function SignOutButton() {
     return <Button onClick={() => signOut()} text='Sign Out' dataCy="signout-button" borderRadius="rounded" color="primary" />
