+++ components/TextField.tsx
@@ -1,13 +1,8 @@
 import { StringUtils } from "@/lib/commons/string_utils";
-import ErrorMessage from "../ErrorMessage";
+import ErrorMessage from "./ErrorMessage";
 import Optional from "./Optional";
 
-type InputFieldType = "text" | "date" | "number";
-export interface InputFieldProps {
-    /**
-     * Input type. Default is 'text'
-     */
-    type?: InputFieldType;
+export interface TextFieldProps {
     /**
      * Input name
      */
@@ -25,14 +20,6 @@ export interface InputFieldProps {
      */
     optional?: boolean;
     /**
-     * Minimum value
-     */
-    min?: number | string;
-    /**
-     * Maximum value
-     */
-    max?: number | string;
-    /**
      * Minimum string length
      */
     minLength?: number;
@@ -41,34 +28,25 @@ export interface InputFieldProps {
      */
     maxLength?: number;
     /**
-     * Default value
-     */
-    defaultValue?: number | string;
-    /**
      * Error message
      */
     errorMessage?: string;
 }
 
-export default function InputField(props: InputFieldProps) {
+export default function TextField(props: TextFieldProps) {
     const hasError = StringUtils.checkInput(props.errorMessage);
 
     return (
         <div className="w-5/6 min-w-full">
             <div className="flex justify-between">
-                <label htmlFor={props.name}>
-                    {props.label}
-                </label>
+                <label htmlFor={props.name}>{props.label}</label>
                 <Optional optional={props.optional} />
             </div>
-            <input type={props.type ?? "text"}
-                placeholder={props.placeholder}
+            <textarea placeholder={props.placeholder}
                 name={props.name}
                 required={!props.optional}
-                min={props.min}
-                max={props.max}
                 minLength={props.minLength}
-                defaultValue={props.defaultValue}
+                maxLength={props.maxLength}
                 className={`mt-2 w-full rounded-md border-2 px-2 py-1 ${hasError ? 'border-red-600' : 'border-slate-200'}`} />
             <ErrorMessage value={props.errorMessage} />
         </div>
